% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_glossaries.R
\name{combine_glossaries}
\alias{combine_glossaries}
\title{Combine glossary sources}
\usage{
combine_glossaries(ropensci_glossary, weblate_glossary)
}
\arguments{
\item{ropensci_glossary}{A standardised rOpenSci glossary data frame.}

\item{weblate_glossary}{A standardised Weblate glossary data frame.}
}
\value{
A combined glossary data frame.
}
\description{
Combines the rOpenSci and Weblate glossaries into a single,
standardised glossary data frame.
}
