% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ropensci_glossary.R
\name{get_ropensci_glossary}
\alias{get_ropensci_glossary}
\title{Get the rOpenSci glossary}
\usage{
get_ropensci_glossary(
  url = paste0("https://raw.githubusercontent.com/",
    "ropensci-review-tools/glossary/master/glossary.csv"),
  retrieved_at = Sys.Date()
)
}
\arguments{
\item{url}{URL of the rOpenSci glossary CSV file.}

\item{retrieved_at}{Date on which the glossary is retrieved.}
}
\value{
A data frame containing the standardised rOpenSci glossary.
}
\description{
Downloads and standardises the rOpenSci glossary for use in the
combined glossary.
}
