% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weblate_glossary.R
\name{get_weblate_glossary}
\alias{get_weblate_glossary}
\title{Get the Weblate glossary}
\usage{
get_weblate_glossary(
  url = paste0("https://translate.rx.studio/download/",
    "r-project/glossary/en/?format=csv"),
  retrieved_at = Sys.Date()
)
}
\arguments{
\item{url}{URL of the Weblate glossary CSV file.}

\item{retrieved_at}{Date on which the glossary is retrieved.}
}
\value{
A data frame containing the standardised Weblate glossary.
}
\description{
Downloads and standardises the Weblate glossary for use in the
combined glossary.
}
