% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\name{glossary}
\alias{glossary}
\title{Display glossary entry}
\usage{
glossary(
  term,
  display = term,
  def = NULL,
  add_to_table = TRUE,
  show = c("term", "def"),
  path = glossary_path()
)
}
\arguments{
\item{term}{The glossary term to link to, can contain spaces}

\item{display}{The display (if different than the term)}

\item{def}{The short definition to display on hover and in the glossary table; if NULL, this will be looked up from the glossary.yml file}

\item{add_to_table}{whether to add to the table created by glossary_table()}

\item{show}{whether to show term with definition on hover (default), or just the term or just the definition}

\item{path}{the path to the glossary file}
}
\value{
string with the link and hover text
}
\description{
Display glossary entry
}
\examples{
glossary("alpha")
glossary("alpha", "$\\\\alpha$")
glossary("alpha", def = "The first letter of the Greek alphabet")
glossary("alpha", show = "term")
glossary("alpha", show = "def")
}
