% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary_add.R
\name{glossary_add}
\alias{glossary_add}
\title{Add a definition to the glossary file}
\usage{
glossary_add(term, definition, path = glossary_path(), replace = FALSE)
}
\arguments{
\item{term}{The term to define}

\item{definition}{The definition to add}

\item{path}{The path of the glossary file}

\item{replace}{Whether to replace an existing definition}
}
\description{
Add a definition to the glossary file
}
