% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\name{glossary_add_to_table}
\alias{glossary_add_to_table}
\title{Add terms and definitions to table}
\usage{
glossary_add_to_table(gloss)
}
\arguments{
\item{gloss}{a named list(term = def)}
}
\value{
NULL; called for side effects
}
\description{
Add terms and definitions to table
}
