% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary_options.R
\name{glossary_persistent}
\alias{glossary_persistent}
\title{Set or get the path to a persistent table object}
\usage{
glossary_persistent(path)
}
\arguments{
\item{path}{the path to the persistent table, or TRUE for the default table name ("glossary-persistent.yml"), or FALSE for no persistence}
}
\value{
path to persistent table object, or FALSE if not set
}
\description{
Quarto books render each chapter in a separate environment, so you need to set a persistent table if you want to display all glossary items in a table in a separate chapter. Set the persistent table at the top of each chapter, after loading the glossary package, and it will automatically add any glossary entries to the persistent table.
}
\examples{
p <- glossary_persistent() # get current path

# set default persistent table path
glossary_persistent(TRUE)

# set non-default path
glossary_persistent(p)
}
