% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary_options.R
\name{glossary_reset}
\alias{glossary_reset}
\title{Reset glossary table}
\usage{
glossary_reset()
}
\value{
NULL; Called for side effects
}
\description{
Resets the list that collects glossary entries for the table.
}
\examples{
glossary_reset()
}
