% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary_style.R
\name{glossary_style}
\alias{glossary_style}
\title{Create CSS styles for glossary entries}
\usage{
glossary_style(
  color = "purple",
  text_decoration = "underline",
  def_bg = "#333",
  def_color = "white"
)
}
\arguments{
\item{color}{Text color of the linked term}

\item{text_decoration}{Style of the linked term; a valid CSS text-decoration string, such as "none", underline" or "red wavy underline"}

\item{def_bg}{Background color of the definition pop-up}

\item{def_color}{Text color of the definition pop-up}
}
\value{
A CSS style string
}
\description{
Set the color and style of the linked in-text terms and pop-up definitions. Colors should be a valid CSS color string, such as "purple" or "#FF0000".
}
\examples{
glossary_style("#003366", "underline")
}
