% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary_table.R
\name{glossary_table}
\alias{glossary_table}
\title{Display glossary table}
\usage{
glossary_table(as_kable = TRUE)
}
\arguments{
\item{as_kable}{if the output should be a kableExtra table or a data frame}
}
\value{
kable table or data frame
}
\description{
All terms defined with \code{\link{glossary}} (since the last call to \code{\link{glossary_reset}}) are added to a list, which this function displays using kable (or outputs as a data frame).
}
\examples{
\donttest{
glossary_reset()
# add a definition to the table
glossary("term", def = "definition", path = NULL)

glossary_table() # show table as kable
glossary_table(FALSE) # or as a data frame
}

}
