% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary_functions.R
\name{reset_glossary}
\alias{reset_glossary}
\title{Reset glossary table}
\usage{
reset_glossary()
}
\description{
Resets the list .myglossary in the parent environment that collects glossary entries for the table
}
\examples{
reset_glossary()

}
