% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary_class.R
\name{glossary}
\alias{glossary}
\title{Glossary class}
\usage{
glossary(definitions_path, glossary_path = "", terms_used = c(),
  header_level = 3)
}
\arguments{
\item{definitions_path}{Where the the definitions of terms are stored. This
is used to show the definitions when hovering over a glossary term in the
text.}

\item{glossary_path}{The file the glossary will be added to. This is used to
link glossary terms in the text to their definitions in the rendered
glossary.}

\item{terms_used}{The terms that will be used. Adding terms to the
constructor (instead of `my_gloss$add("new term")`) will include them as if
they were added with `my_gloss$add()`.}

\item{header_level}{How big the headers are for each term in the rendered
glossary. Larger numbers mean smaller titles.}
}
\value{
An `R6Class` object of class `Glossary`
}
\description{
This is used to add terms to a glossary
}
\examples{
\dontrun{
my_gloss <- glossary()
}

}
