% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary_options.R
\name{glossary_options}
\alias{glossary_options}
\title{Set/get global glossary options}
\usage{
glossary_options(...)
}
\arguments{
\item{...}{One of three: (1) nothing, then returns all options as a list; (2) a name of an option element, then returns its value; (3) a name-value pair which sets the corresponding option to the new value (and returns nothing)}
}
\value{
a list of options, values of an option, or nothing
}
\description{
Global glossary options are used, for example, to set the default path for the glossary file.
}
\keyword{internal}
