% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary_options.R
\name{glossary_path}
\alias{glossary_path}
\title{Set or get the default glossary path}
\usage{
glossary_path(path = NULL)
}
\arguments{
\item{path}{the path to the glossary file}
}
\value{
path string if path is NULL
}
\description{
Set or get the default glossary path
}
\examples{
glossary_path() # get current path
}
