% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary_options.R
\name{glossary_popup}
\alias{glossary_popup}
\title{Set or get the default popup type}
\usage{
glossary_popup(popup = NULL)
}
\arguments{
\item{popup}{If Null, get the current default popup type, or set to one of "click", "hover", or "none"}
}
\value{
string if popup is NULL
}
\description{
Set or get the default popup type
}
\examples{
glossary_popup() # get current popup style
}
