% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary_style.R
\name{glossary_style}
\alias{glossary_style}
\title{Create CSS styles for glossary entries}
\usage{
glossary_style(color = "purple", text_decoration = "none")
}
\arguments{
\item{color}{A valid CSS colour string, such as "purple" or "#FF0000"}

\item{text_decoration}{A valid CSS text-decoration string, such as "underline" or "red wavy underline"}
}
\value{
A CSS style string
}
\description{
Create CSS styles for glossary entries
}
\examples{
glossary_style("#003366", "underline")
}
