% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary_table.R
\name{glossary_table}
\alias{glossary_table}
\title{Display glossary table}
\usage{
glossary_table(as_kable = TRUE)
}
\arguments{
\item{as_kable}{if the output should be a kableExtra table or a data frame}
}
\value{
kable table or data frame
}
\description{
Display glossary table
}
