% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary_functions.R
\name{glossary_table}
\alias{glossary_table}
\title{Display glossary table}
\usage{
glossary_table(link = TRUE, as_kable = TRUE)
}
\arguments{
\item{link}{whether to include a link to https://psyteachr.github.io/glossary/}

\item{as_kable}{if the output should be a kable table or a data table}
}
\value{
kable table or data table
}
\description{
Display glossary table
}
\examples{
glossary("alpha")
glossary("beta")
glossary_table()
}
